feat: update composer dependencies to support Laravel 11
This commit updates the composer.json file to include the latest versions of the `illuminate/database`, `illuminate/http`, `illuminate/pagination`, and `illuminate/validation` packages. The new versions specified are "^9.35||^10.11||^11.0", which are compatible with Laravel 11.

These updates ensure that the project can take advantage of the latest features and improvements in the Laravel framework.

Actualizar dependencias de composer para soportar Laravel 11